+++ src/TestRunner/Test.ts
@@ -1,9 +1,9 @@
 import { ApiPromise, WsProvider } from '@polkadot/api';
 import { KeyringPair } from '@polkadot/keyring/types';
+import Mocha from 'mocha';
 import path from 'path';
 import { generateAccounts } from '../Accounts';
 import { HalvaTestConfig } from "./Config/HalvaTestConfig";
-import { CreateMocha } from './TestRunner';
  // tslint:disable: variable-name
 declare global {
     var halva_polkadot: ApiPromise;
@@ -18,7 +18,7 @@ export const Run = async (config: HalvaTestConfig) => {
       const provider = new WsProvider(config.network.test.ws);
       const polkadot = await ApiPromise.create({ provider });
       const accounts = await generateAccounts(10, config.network.test.mnemonic);
-      const mocha = CreateMocha(config);
+      const mocha = this.CreateMocha(config);
       config.testingFiles.forEach(file => {
         mocha.addFile(file);
       });
@@ -33,5 +33,23 @@ export const SetTestGlobal = (accounts: KeyringPair[], polkadot: ApiPromise) => 
     globalThis.halva_polkadot = polkadot;
 }
 
-Run({testingFiles: ['/home/staler/halva/src/tests/example.test.js'], network: { test: {ws: "ws://127.0.0.1:9944",
+
+export const CreateMocha = (config: HalvaTestConfig): Mocha => {
+    const mochaConfig = config.mocha || {};
+ // Propagate --bail option to mocha
+    mochaConfig.bail = config.bail;
+
+    // If the command line overrides color usage, use that.
+    if (config.colors != null) { mochaConfig.useColors = config.colors; }
+    // Default to true if configuration isn't set anywhere.
+    if (mochaConfig.useColors == null) {
+     mochaConfig.useColors = true;
+    }
+
+    const mocha = new Mocha(mochaConfig);
+
+    return mocha;
+}
+
+Run({testingFiles: ['./src/tests/example.test.js'], network: { test: {ws: "ws://127.0.0.1:9944",
 mnemonic: "bottom drive obey lake curtain smoke basket hold race lonely fit walk"}}, bail: false, mocha: {}});
